Chore: Move lib/logging to lib/util/logging
