+++ lib/util/logging.js
