+++ tests/lib/cli.js
@@ -37,7 +37,7 @@ describe("cli", () => {
         error: sinon.spy()
     };
     const cli = proxyquire("../../lib/cli", {
-        "./logging": log
+        "./util/logging": log
     });
 
     /**
@@ -58,7 +58,7 @@ describe("cli", () => {
 
         const localCLI = proxyquire("../../lib/cli", {
             "./cli-engine": fakeCLIEngine,
-            "./logging": log
+            "./util/logging": log
         });
 
         localCLI.execute(cmd);
@@ -697,7 +697,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             localCLI.execute("--no-inline-config .");
@@ -719,7 +719,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute(".");
@@ -755,7 +755,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix .");
@@ -791,7 +791,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix .");
@@ -828,7 +828,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix --quiet .");
@@ -844,7 +844,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix .", "foo = bar;");
@@ -878,7 +878,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix-dry-run .");
@@ -914,7 +914,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix-dry-run .");
@@ -951,7 +951,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix-dry-run --quiet .");
@@ -987,7 +987,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix-dry-run .", "foo = bar;");
@@ -1002,7 +1002,7 @@ describe("cli", () => {
 
             localCLI = proxyquire("../../lib/cli", {
                 "./cli-engine": fakeCLIEngine,
-                "./logging": log
+                "./util/logging": log
             });
 
             const exitCode = localCLI.execute("--fix --fix-dry-run .", "foo = bar;");
